m100pfsevp: remove the default eth mac address settings
Like for the other boards from ARIES Embedded, we use the env variables
"ethaddr" and "eth1addr" to define the mac address for the two Ethernet
interfaces at commissioning time. For the user's convenience we also
enable random mac addresses in case "ethaddr" or "eth1addr" are unset.

For the Icicle, the last three bytes of the mac address was derived
from the serial number from the System Services Block. We keep and
print it out to the console.

Signed-off-by: Wolfgang Grandegger <wg@grandegger.com>