lmb: allow subset regions with different flags
If the kernel's DT contains reserved memory regions that are a subset of
memory regions known about by U-Boot, U-Boot will fail to reserve them.
This happens as subset regions with different flags to their super-sets
hit a failure case in the lmb code.
For us, this manifests while reserving a no-map region that is subset to
a normal memory region known about by U-Boot.

This is a naive change, and something more involved (likely including
reverting the recent changes to coalescing regions) will be needed to
complete this fix.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>